// src/data/photos.generated.ts
export type StudentPhoto = {
  slug: string;
  src: string;
  srcSmall: string;
  width: number;
  height: number;
  car: 'tesla' | 'sonata';
  blur: string;
};

export const STUDENT_PHOTOS: readonly StudentPhoto[] = [
  {
    "slug": "aileen",
    "src": "/photos/aileen-1100.webp",
    "srcSmall": "/photos/aileen-640.webp",
    "width": 3024,
    "height": 4032,
    "car": "tesla",
    "blur": "data:image/webp;base64,UklGRo4AAABXRUJQVlA4IIIAAAAQBQCdASoUABsAPxF+slOsKCQiqAqpgCIJQBBOAFJ82vbVquCLjlDXIP7CBMXOucAAAP5fzAaaKiJQTYSjOXw0QJRk/0CL9+cNK8N8+zMS/83J1OEQ7fRLuHdf++T4FVUxlMSU6wGn5KPH49X1Lteneh/1btf+jYzkFf4G0ZuvAAAA"
  },
  {
    "slug": "img-4037",
    "src": "/photos/img-4037-1100.webp",
    "srcSmall": "/photos/img-4037-640.webp",
    "width": 1512,
    "height": 2016,
    "car": "tesla",
    "blur": "data:image/webp;base64,UklGRoIAAABXRUJQVlA4IHYAAADQBACdASoUABsAPxF2sVGsJySiqA1RgCIJYgC06AeSvc7ooS7RcoSILhBglIhYAADSezoAUkR5/zSGsguA/JvrUdHVU8ejjPxwzsElIEis70ylwGB9zMCy6wM3o1DX0r7yHbeQdS03RLMVpY3wjULuux2JSAAA"
  },
  {
    "slug": "img-4044",
    "src": "/photos/img-4044-1100.webp",
    "srcSmall": "/photos/img-4044-640.webp",
    "width": 1512,
    "height": 2016,
    "car": "tesla",
    "blur": "data:image/webp;base64,UklGRnwAAABXRUJQVlA4IHAAAADwBACdASoUABsAPxGAuFWsKCUjKAgBgCIJZAC06AuKl/vtH37fm8RSej1/QiqKrAAA9b/xhyxBqnmNtW81ufNcbfW92uNzB6tyfl6GaJup9kWcee04gGP3qJ9id4csUDq0ZEIviGbBdRhc7Jzri4AA"
  },
  {
    "slug": "img-4436-1",
    "src": "/photos/img-4436-1-1100.webp",
    "srcSmall": "/photos/img-4436-1-640.webp",
    "width": 3024,
    "height": 4032,
    "car": "tesla",
    "blur": "data:image/webp;base64,UklGRm4AAABXRUJQVlA4IGIAAAAQBACdASoUABsAPxFytFMsJiUisBgIAYAiCWMAygAI8MSTyd/zqRngAAD2TBWwusOtRCKzVf4W6DwV5tB8y+KcGHl9QiR/ubI6fccNLdrHHbs8RezsVd/KalUYhD6v6AAAAA=="
  },
  {
    "slug": "img-4810",
    "src": "/photos/img-4810-1100.webp",
    "srcSmall": "/photos/img-4810-640.webp",
    "width": 3024,
    "height": 4032,
    "car": "tesla",
    "blur": "data:image/webp;base64,UklGRnIAAABXRUJQVlA4IGYAAAAwBQCdASoUABsAPxFwsFAsJiSisBgMAYAiCWUArAAQ6yL3yO/7QfCYgGkNUA5qRs6rQAD+rwShBZkm8JxRvPuFRAWaNehjfSpMkj5LYjAxYoBSN1aXIRrt8uh87zg8qhPQjVxSAAA="
  },
  {
    "slug": "img-5238",
    "src": "/photos/img-5238-1100.webp",
    "srcSmall": "/photos/img-5238-640.webp",
    "width": 3024,
    "height": 4032,
    "car": "tesla",
    "blur": "data:image/webp;base64,UklGRngAAABXRUJQVlA4IGwAAABQBACdASoUABsAPxF4s1IsJySiqA1RgCIJZwDLLA3CD7y1zu3z7AoGc7IAAP7V5GzoIVhcLBY8qS7dMqXm6cU96bPk06oVI8DvdNuXOqniZLtAicHGHMdOV/Tyswnjlt3eZlkGV6xrHGoAAAA="
  },
  {
    "slug": "img-5708",
    "src": "/photos/img-5708-1100.webp",
    "srcSmall": "/photos/img-5708-640.webp",
    "width": 1512,
    "height": 2016,
    "car": "tesla",
    "blur": "data:image/webp;base64,UklGRnYAAABXRUJQVlA4IGoAAACwBACdASoUABsAPv1sq1ArJqQit/qoAWAfiUAYZoEk0TxyBrztr9b1gsZ2MmoAAPofivDvZgm5ZA7MZN1dAyju8VzZzWeQoAI2bV4GEBJX8X4+qjxRfHfCT1Qt/jiaDglxr7cpNx7gAAAA"
  },
  {
    "slug": "img-5773",
    "src": "/photos/img-5773-1100.webp",
    "srcSmall": "/photos/img-5773-640.webp",
    "width": 1512,
    "height": 2016,
    "car": "tesla",
    "blur": "data:image/webp;base64,UklGRn4AAABXRUJQVlA4IHIAAACwBACdASoUABsAPxF6t1MsJ6WiqAqpgCIJQBllgfBr/sAe4g7qWJQCeaJMau1QAPyDBjBH80XHZfjfaUzQaNXMMSnwpBiutEFk6zd93X+LRzb4EqBg8VwhhYs7+/I6KnBYqNUxB68zw+YLVGEJwBmAAAA="
  },
  {
    "slug": "img-5808",
    "src": "/photos/img-5808-1100.webp",
    "srcSmall": "/photos/img-5808-640.webp",
    "width": 1512,
    "height": 2016,
    "car": "tesla",
    "blur": "data:image/webp;base64,UklGRn4AAABXRUJQVlA4IHIAAAAQBQCdASoUABsAPxF6sFKsJ6QiqA1RgCIJZACxH0LA4Cj+wdcx0fdAKR6F1HPKTPdAAP0cEqTwTZZS9nfGI5sr+TfSIi64olVtf5xRjrmcwmVRUTCPx67oGfw+qa6NhBzZ6/CYUm2beiohiF3S/b32AAA="
  },
  {
    "slug": "vera-tesla",
    "src": "/photos/vera-tesla-1100.webp",
    "srcSmall": "/photos/vera-tesla-640.webp",
    "width": 1512,
    "height": 2016,
    "car": "tesla",
    "blur": "data:image/webp;base64,UklGRnYAAABXRUJQVlA4IGoAAACwBACdASoUABsAPxF4sVIsJySiqA1RgCIJZQDDcA0xJPlLCdv6Cse8aDWVIyrMAP6lFk3KOtmlj88fRe3NYIRLCHdpGbXWDL38FNzE5BD+KvC9asztq9+wLtYotRE4xGq5mU+BR/YL9QAA"
  }
];